src/server/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
-e git+ssh://git@github.com/Miloye1/doc-2-geepeetee-service.git@1db98f8260708e906ed22c94e3ae4c50048fa169#egg=geepeetee&subdirectory=src/modules
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.3
Werkzeug==3.0.2
